GEN_AI/Design & Build a RAG-Based Customer Support Assistant (with LangGraph & HITL)/requirements.txt
langchain==0.3.7
langchain-community==0.3.5
langchain-openai==0.2.8
langgraph==0.2.43
chromadb==0.5.15
pypdf==5.0.1
sentence-transformers==3.2.1
python-dotenv==1.0.1
